requirements_raspi.txt
altgraph==0.17.4
bleak==0.22.3
bluezero==0.9.0
dbus-fast==2.24.3
dbus-python==1.3.2
glfw==2.7.0
numpy==1.26.4
packaging==24.0
pillow==11.0.0
pycairo==1.27.0
PyGObject==3.50.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
PyOpenGL==3.1.7
typing_extensions==4.12.2 